+++ sway-lsp/src/tests/lsp.rs
@@ -1,5 +1,5 @@
 use crate::server::{self, Backend};
-use crate::utils::test::{
+use crate::tests::utils::{
     assert_server_requests, dir_contains_forc_manifest, doc_comments_dir, e2e_language_dir,
     e2e_test_dir, get_fixture, runnables_test_dir, sway_workspace_dir, test_fixtures_dir,
 };
