+++ sway-lsp/src/utils/mod.rs
@@ -3,5 +3,3 @@ pub mod debug;
 pub(crate) mod keyword_docs;
 pub mod markdown;
 pub(crate) mod markup;
-#[cfg(test)]
-pub(crate) mod test;
